✨ Implementada corrida com metade dos pontos

+++ Formula1.Data/Entities/EquipeTemporada.cs
@@ -10,7 +10,7 @@
 
         public int Temporada { get; set; }
 
-        public int Pontos { get; set; }
+        public double Pontos { get; set; }
 
         public int Posicao { get; set; }
     }
