+++ Formula1.Data/Entities/PilotoTemporada.cs
@@ -6,7 +6,7 @@
 
         public int Temporada { get; set; }
 
-        public int Pontos { get; set; }
+        public double Pontos { get; set; }
 
         public int Posicao { get; set; }
     }
