+++ Formula1.Data/Models/EquipeTemporadaInclusao.cs
@@ -6,11 +6,11 @@
 
         public int Temporada { get; set; }
 
-        public int Pontos { get; set; }
+        public double Pontos { get; set; }
 
         public int Posicao { get; set; }
 
-        public EquipeTemporadaInclusao(int equipeId, int temporada, int pontos, int posicao)
+        public EquipeTemporadaInclusao(int equipeId, int temporada, double pontos, int posicao)
         {
             EquipeId = equipeId;
             Temporada = temporada;
