+++ Formula1.Domain/Services/GraficoCampeonatoEquipesService.cs
@@ -21,9 +21,9 @@ namespace Formula1.Domain.Services
 
             foreach (var equipe in campeonato.Equipes)
             {
-                var pontos = new int?[campeonato.Corridas.Count];
+                var pontos = new double?[campeonato.Corridas.Count];
 
-                int soma = 0;
+                double soma = 0;
                 foreach (var corrida in campeonato.Corridas)
                 {
                     if (corrida.Resultados.Count == 0)
